fix bug for retrieve  CalculateRouteResponse Update input call Update model

+++ src/main/java/com/heromaps/client/dto/WaypointDTO.java
@@ -2,7 +2,7 @@ package com.heromaps.client.dto;
 
 public class WaypointDTO {
 
-	public Float lat;
-	public Float lng;
+	public Double lat;
+	public Double lng;
 
 }
