+++ src/main/java/com/heromaps/client/result/SegmentsResponse.java
@@ -2,58 +2,58 @@ package com.heromaps.client.result;
 
 public class SegmentsResponse {
 
-	private float startinglat;
-	private float startinglng;
-	private float endinglat;
-	private float endinglnf;
-	private float distance;
-	private long time;
+	private double startinglat;
+	private double startinglng;
+	private double endinglat;
+	private double endinglnf;
+	private double distance;
+	private double time;
 
-	public float getStartinglat() {
+	public double getStartinglat() {
 		return startinglat;
 	}
 
-	public void setStartinglat(float startinglat) {
+	public void setStartinglat(double startinglat) {
 		this.startinglat = startinglat;
 	}
 
-	public float getStartinglng() {
+	public double getStartinglng() {
 		return startinglng;
 	}
 
-	public void setStartinglng(float startinglng) {
+	public void setStartinglng(double startinglng) {
 		this.startinglng = startinglng;
 	}
 
-	public float getEndinglat() {
+	public double getEndinglat() {
 		return endinglat;
 	}
 
-	public void setEndinglat(float endinglat) {
+	public void setEndinglat(double endinglat) {
 		this.endinglat = endinglat;
 	}
 
-	public float getEndinglnf() {
+	public double getEndinglnf() {
 		return endinglnf;
 	}
 
-	public void setEndinglnf(float endinglnf) {
+	public void setEndinglnf(double endinglnf) {
 		this.endinglnf = endinglnf;
 	}
 
-	public float getDistance() {
+	public double getDistance() {
 		return distance;
 	}
 
-	public void setDistance(float distance) {
+	public void setDistance(double distance) {
 		this.distance = distance;
 	}
 
-	public long getTime() {
+	public double getTime() {
 		return time;
 	}
 
-	public void setTime(long time) {
+	public void setTime(double time) {
 		this.time = time;
 	}
